New survey update
Comments and update status question. After you take the new survey once, it asks to "update status" rather than just asking for the status. Fixed the clicker for this question (might not work for first-time survey takers now) and added comments that explain connect each driver action to a question.